Fixing collider's + enemy
Fixing most of the collider's but and adding an enemy to the game,
though there are still some minor issues with the collider with above the player blocks it's working and those bugs will never be encountered.
Need to add behavior on collision with the player. Adding a lifespan rendering queue to avoir memory issues